compose/django/requirements.txt
anyio==3.6.2
ari==0.1.3
asgiref==3.5.2
asks==3.0.0
async-generator==1.10
asyncari==0.10.8
asyncswagger11==0.12.4
asyncwebsockets==0.9.4
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Django==4.0.3
django-environ==0.9.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
multidict==6.0.2
outcome==1.2.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pystrix==1.1.8
pytz==2022.6
rel==0.4.9
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
swaggerpy==0.2.1
trio==0.22.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.12
websocket-client==1.4.2
wsproto==1.2.0
yarl==1.8.1
zope.interface==5.5.1
